bug fixed in .deposit()
As description is an optional argument, I've put * to make it optional. 
However, because of this, description is by default a tuple (I don't know why).
In this update I fixed the problem of tuple, now description is a str in deposit.
But still not fixed in withdraw. It's still a tuple.
How creepy.